Use a threshold above which switch from a bounded to an unbounded priority queue.
In order to process ordered queries we use a PriorityQueue to store the rows
and then pop them in the correct order. This caused issues as before we used
a bounded queue with the bound being limit + offset. This was overly pessimistic
as the number of rows available in the table could be much lower than that,
but the PriorityQueue container was reserving the memory for this worse case.

This changes the implementation to use a threshold over which we switch from
using a bounded queue to an unbounded queue that will grow as needed.

(cherry picked from commit 455570b8988cd79070c8071347b31da80850ade1)

# Conflicts:
#	blackbox/docs/appendices/release-notes/unreleased.rst